@@ -11,9 +11,12 @@ export 'src/widgets/gesture_detector.dart';
 export 'src/widgets/icon.dart';
 export 'src/widgets/icon_data.dart';
 export 'src/widgets/image.dart';
+export 'src/widgets/navigator.dart';
+export 'src/widgets/overlay.dart';
+export 'src/widgets/pages.dart';
+export 'src/widgets/routes.dart';
 export 'src/widgets/spacer.dart';
 export 'src/widgets/scroll_controller.dart';
 export 'src/widgets/scroll_view.dart';
 export 'src/widgets/text.dart';
-export 'src/widgets/navigator.dart';
 ''';
